third_party/WebKit/LayoutTests/fast/dom/DOMException/prototype-object.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>

description("This tests the prototype chain of DOMException objects.")

var e;
try {
  document.appendChild(document);
   // raises a HIERARCHY_REQUEST_ERR
} catch (err) {
  e = err;
}

shouldBeEqualToString("e.toString()", "HierarchyRequestError: Failed to execute 'appendChild' on 'Node': The new child element contains the parent.");
shouldBeEqualToString("Object.prototype.toString.call(e)", "[object DOMException]");
shouldBeEqualToString("Object.prototype.toString.call(e.__proto__)", "[object DOMExceptionPrototype]");
shouldBeEqualToString("Object.prototype.toString.call(e.__proto__.__proto__)", "[object Error]");
shouldBeEqualToString("e.constructor.toString()", "function DOMException() { [native code] }");
shouldBeTrue("e instanceof DOMException");
shouldBeTrue("e instanceof Error");
shouldBe("e.constructor", "window.DOMException");
shouldBe("e.HIERARCHY_REQUEST_ERR", "e.constructor.HIERARCHY_REQUEST_ERR");
shouldBe("e.HIERARCHY_REQUEST_ERR", "3");
shouldBe("e.code", "3");
shouldBeEqualToString("e.name", "HierarchyRequestError");
shouldBeEqualToString("e.message", "Failed to execute 'appendChild' on 'Node': The new child element contains the parent.");

</script>
<script src="../../js/resources/js-test-post.js"></script>
